# libroach: port hash/crc32's ieee implementation to c