--- a/c-deps/libroach/encoding.cc
+++ b/c-deps/libroach/encoding.cc
@@ -15,10 +15,7 @@
 // Author: Spencer Kimball (spencer.kimball@gmail.com)
 // Author: Peter Mattis (peter@cockroachlabs.com)
 
-#include "rocksdb/slice.h"
 #include "encoding.h"
-
-// TODO(benesch): Set up a CI pipeline to test these functions.
 
 void EncodeUint32(std::string* buf, uint32_t v) {
   const uint8_t tmp[sizeof(v)] = {