This is a Friday project to validate a hunch I have that moving all of restore's SST work into C++ will yield a nice performance boost. Getting a fast CRC32 implementation was step 1. Please don't give this a thorough review yet—I haven't had the chance to ask @danhhz if this is even a road we want to seriously explore.

Only the last commit here is out for (proof-of-concept only) review; the first two commits are tracked in #17035 and #17039 respectively.

---

Port Go's highly-optimized AMD64 implementation of CRC32 with the IEEE polynomial to C.

Go 1.9beta2:
```
BenchmarkCRC32/poly=IEEE/size=15/align=0-8         	30000000	        52.0 ns/op	 288.37 MB/s
BenchmarkCRC32/poly=IEEE/size=15/align=1-8         	30000000	        51.9 ns/op	 289.20 MB/s
BenchmarkCRC32/poly=IEEE/size=40/align=0-8         	30000000	        47.5 ns/op	 842.13 MB/s
BenchmarkCRC32/poly=IEEE/size=40/align=1-8         	30000000	        47.5 ns/op	 841.89 MB/s
BenchmarkCRC32/poly=IEEE/size=512/align=0-8        	20000000	        66.3 ns/op	7721.03 MB/s
BenchmarkCRC32/poly=IEEE/size=512/align=1-8        	20000000	        66.7 ns/op	7679.31 MB/s
BenchmarkCRC32/poly=IEEE/size=1kB/align=0-8        	20000000	       111 ns/op	9217.26 MB/s
BenchmarkCRC32/poly=IEEE/size=1kB/align=1-8        	20000000	       115 ns/op	8840.18 MB/s
BenchmarkCRC32/poly=IEEE/size=4kB/align=0-8        	 5000000	       392 ns/op	10443.18 MB/s
BenchmarkCRC32/poly=IEEE/size=4kB/align=1-8        	 5000000	       342 ns/op	11969.82 MB/s
BenchmarkCRC32/poly=IEEE/size=32kB/align=0-8       	  500000	      2574 ns/op	12730.15 MB/s
BenchmarkCRC32/poly=IEEE/size=32kB/align=1-8       	  500000	      2599 ns/op	12603.22 MB/s
```

C:

```
BenchmarkCRC32/poly=IEEE/size=15/align=0-8		 30000000	23.5 ns/op	608.08 MB/s
BenchmarkCRC32/poly=IEEE/size=15/align=1-8		 30000000	23.0 ns/op	621.36 MB/s
BenchmarkCRC32/poly=IEEE/size=40/align=0-8		 30000000	37.4 ns/op	1018.66 MB/s
BenchmarkCRC32/poly=IEEE/size=40/align=1-8		 30000000	38.0 ns/op	1002.82 MB/s
BenchmarkCRC32/poly=IEEE/size=512/align=0-8		 20000000	51.1 ns/op	9548.67 MB/s
BenchmarkCRC32/poly=IEEE/size=512/align=1-8		 20000000	50.7 ns/op	9637.17 MB/s
BenchmarkCRC32/poly=IEEE/size=1kB/align=0-8		 20000000	91.3 ns/op	10698.23 MB/s
BenchmarkCRC32/poly=IEEE/size=1kB/align=1-8		 20000000	91.9 ns/op	10625.31 MB/s
BenchmarkCRC32/poly=IEEE/size=4kB/align=0-8		  5000000	332.9 ns/op	11735.48 MB/s
BenchmarkCRC32/poly=IEEE/size=4kB/align=1-8		  5000000	332.3 ns/op	11754.54 MB/s
BenchmarkCRC32/poly=IEEE/size=32kB/align=0-8	   500000	2516.3 ns/op	12418.92 MB/s
BenchmarkCRC32/poly=IEEE/size=32kB/align=1-8	   500000	2624.0 ns/op	11909.40 MB/s
```

benchstat:

```
name                                 old time/op    new time/op     delta
CRC32/poly=IEEE/size=15/align=0-8      53.0ns ± 1%     23.9ns ± 3%   -54.83%  (p=0.008 n=5+5)
CRC32/poly=IEEE/size=15/align=1-8      53.1ns ± 1%     23.7ns ± 3%   -55.31%  (p=0.008 n=5+5)
CRC32/poly=IEEE/size=40/align=0-8      48.2ns ± 1%     28.7ns ± 2%   -40.42%  (p=0.008 n=5+5)
CRC32/poly=IEEE/size=40/align=1-8      48.5ns ± 2%     28.9ns ± 2%   -40.32%  (p=0.008 n=5+5)
CRC32/poly=IEEE/size=512/align=0-8     65.1ns ± 2%     52.0ns ± 1%   -20.21%  (p=0.008 n=5+5)
CRC32/poly=IEEE/size=512/align=1-8     65.7ns ± 1%     52.6ns ± 2%   -19.91%  (p=0.008 n=5+5)
CRC32/poly=IEEE/size=1kB/align=0-8      106ns ± 2%       94ns ± 1%   -11.40%  (p=0.016 n=5+4)
CRC32/poly=IEEE/size=1kB/align=1-8      106ns ± 1%       97ns ± 8%    -8.96%  (p=0.008 n=5+5)
CRC32/poly=IEEE/size=4kB/align=0-8      357ns ± 9%      341ns ± 2%    -4.49%  (p=0.008 n=5+5)
CRC32/poly=IEEE/size=4kB/align=1-8      347ns ± 1%      338ns ± 2%      ~     (p=0.056 n=5+5)
CRC32/poly=IEEE/size=32kB/align=0-8    2.59µs ± 1%     2.63µs ± 5%      ~     (p=0.690 n=5+5)
CRC32/poly=IEEE/size=32kB/align=1-8    2.64µs ± 1%     2.67µs ±13%      ~     (p=0.500 n=5+5)

name                                 old speed      new speed       delta
CRC32/poly=IEEE/size=15/align=0-8     283MB/s ± 1%    598MB/s ± 3%  +111.25%  (p=0.008 n=5+5)
CRC32/poly=IEEE/size=15/align=1-8     282MB/s ± 1%    603MB/s ± 3%  +113.60%  (p=0.008 n=5+5)
CRC32/poly=IEEE/size=40/align=0-8     829MB/s ± 1%   1328MB/s ± 2%   +60.19%  (p=0.008 n=5+5)
CRC32/poly=IEEE/size=40/align=1-8     826MB/s ± 2%   1320MB/s ± 2%   +59.89%  (p=0.008 n=5+5)
CRC32/poly=IEEE/size=512/align=0-8   7.86GB/s ± 2%   9.40GB/s ± 1%   +19.52%  (p=0.008 n=5+5)
CRC32/poly=IEEE/size=512/align=1-8   7.79GB/s ± 1%   9.28GB/s ± 2%   +19.12%  (p=0.008 n=5+5)
CRC32/poly=IEEE/size=1kB/align=0-8   9.58GB/s ± 1%  10.36GB/s ± 0%    +8.14%  (p=0.016 n=5+4)
CRC32/poly=IEEE/size=1kB/align=1-8   9.60GB/s ± 1%  10.12GB/s ± 8%      ~     (p=0.151 n=5+5)
CRC32/poly=IEEE/size=4kB/align=0-8   11.5GB/s ± 8%   11.5GB/s ± 2%      ~     (p=0.310 n=5+5)
CRC32/poly=IEEE/size=4kB/align=1-8   11.8GB/s ± 1%   11.6GB/s ± 3%      ~     (p=0.310 n=5+5)
CRC32/poly=IEEE/size=32kB/align=0-8  12.6GB/s ± 1%   11.9GB/s ± 5%    -5.76%  (p=0.008 n=5+5)
CRC32/poly=IEEE/size=32kB/align=1-8  12.4GB/s ± 1%   11.8GB/s ±12%    -5.44%  (p=0.008 n=5+5)
```