@bdarnell 2017-07-17
RocksDB already has a crc32 implementation. Could we use that one (looks like we'd need to export the header) instead of introducing a new one?