@danhhz 2017-07-17
@bdarnell it's the wrong crc32. They have Castagnoli. If we use this, it would be for verifying our individual kv checksums, which are IEEE