@danhhz 2017-07-17
Cool stuff, you'll definitely want someone that speaks c++ to do the review.

I'd be interested to see how this affects the AddSSTable benchmarks. Given this, it should be pretty easy to move the current impl of verifySSTable entirely to c++ and see what happens to BenchmarkAddSSTable and BenchmarkImport/AddSSTable