@petermattis 2017-07-17
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/> modulo using curly braces always.

@raduberinde Can you take a look at the crc stuff given your familiarity with the Go crc code?

---

Review status: 0 of 51 files reviewed at latest revision, 3 unresolved discussions, some commit checks failed.

---

*[c-deps/libroach/LICENSE.go, line 1 at r3](https://reviewable.io:443/reviews/cockroachdb/cockroach/17040#-KpGiAabPrhs7jAwVBUj:-KpGiAabPrhs7jAwVBUk:bewioin) ([raw file](https://github.com/cockroachdb/cockroach/blob/eb11d984700e79ebb9edfb1b35b61bbf7e7ffed2/c-deps/libroach/LICENSE.go#L1)):*
> ```Go
> Copyright (c) 2009 The Go Authors. All rights reserved.
> ```

Giving this file the `.go` suffix is a bit confusing since that makes it appear to be a Go source file. Perhaps `LICENSE.golang`. 

---

*[c-deps/libroach/testutil.c, line 24 at r3](https://reviewable.io:443/reviews/cockroachdb/cockroach/17040#-KpGiO6Mp7HjPtCE1SGR:-KpGiO6Mp7HjPtCE1SGS:bacw25m) ([raw file](https://github.com/cockroachdb/cockroach/blob/eb11d984700e79ebb9edfb1b35b61bbf7e7ffed2/c-deps/libroach/testutil.c#L24)):*
> ```C
> make_bytes(size_t len) {
> 	long* buf = malloc(len);
> 	if (buf == NULL)
> ```

Almost all C/C++ style guidelines recommend using curly braces even when they aren't necessary. For example, `if (buf == NULL) { err(...); }`

Also, you seem to be using tabs for indentation in this file while other files use spaces. I'd stick with spaces.

---

*[c-deps/libroach/testutil.h, line 17 at r3](https://reviewable.io:443/reviews/cockroachdb/cockroach/17040#-KpGj0XzyoRGgP8udweb:-KpGj0XzyoRGgP8udwec:ba7t5w) ([raw file](https://github.com/cockroachdb/cockroach/blob/eb11d984700e79ebb9edfb1b35b61bbf7e7ffed2/c-deps/libroach/testutil.h#L17)):*
> ```objc
> // Author: Nikhil Benesch (nikhil.benesch@gmail.com)
> 
> #define ARRAY_SIZE(a) (sizeof(a) / sizeof(a[0]))
> ```

You only seem to use this in `crc32_bench.c`. Move the definition there?

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/17040#-:-KpGjIRSM2eYjU2-AaN-:bpv61da)*
<!-- Sent from Reviewable.io -->