@tamird 2017-07-18
I'd like to review this - could you ping when it's rebased on top of #17035?

---

Review status: 0 of 51 files reviewed at latest revision, 3 unresolved discussions, some commit checks failed.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/17040#-:-KpI5BzrsDLbqmCGF0LF:b-qndbe8)*
<!-- Sent from Reviewable.io -->