@benesch 2017-07-18
Ping!

Rebuilding the builder now.

---

Review status: 0 of 15 files reviewed at latest revision, 3 unresolved discussions.

---

*[c-deps/libroach/testutil.c, line 24 at r3](https://reviewable.io:443/reviews/cockroachdb/cockroach/17040#-KpGiO6Mp7HjPtCE1SGR:-KpI90zGyWe9mnKWuayJ:b-glx2gt) ([raw file](https://github.com/cockroachdb/cockroach/blob/eb11d984700e79ebb9edfb1b35b61bbf7e7ffed2/c-deps/libroach/testutil.c#L24)):*
<details><summary><i>Previously, petermattis (Peter Mattis) wrote…</i></summary><blockquote>

Almost all C/C++ style guidelines recommend using curly braces even when they aren't necessary. For example, `if (buf == NULL) { err(...); }`

Also, you seem to be using tabs for indentation in this file while other files use spaces. I'd stick with spaces.
</blockquote></details>

Yeah, it's a bad habit I've picked up, and there's no go fmt to tell me I can't. Done and done.

---

*[c-deps/libroach/testutil.h, line 17 at r3](https://reviewable.io:443/reviews/cockroachdb/cockroach/17040#-KpGj0XzyoRGgP8udweb:-KpI97v0pTyoW_POsigk:be74cn7) ([raw file](https://github.com/cockroachdb/cockroach/blob/eb11d984700e79ebb9edfb1b35b61bbf7e7ffed2/c-deps/libroach/testutil.h#L17)):*
<details><summary><i>Previously, petermattis (Peter Mattis) wrote…</i></summary><blockquote>

You only seem to use this in `crc32_bench.c`. Move the definition there?
</blockquote></details>

It's used in crc32_test.c as well.

---

*[c-deps/libroach/LICENSE.go, line 1 at r3](https://reviewable.io:443/reviews/cockroachdb/cockroach/17040#-KpGiAabPrhs7jAwVBUj:-KpI9VBz4n42x2XcsmSU:bombmv2) ([raw file](https://github.com/cockroachdb/cockroach/blob/eb11d984700e79ebb9edfb1b35b61bbf7e7ffed2/c-deps/libroach/LICENSE.go#L1)):*
<details><summary><i>Previously, petermattis (Peter Mattis) wrote…</i></summary><blockquote>

Giving this file the `.go` suffix is a bit confusing since that makes it appear to be a Go source file. Perhaps `LICENSE.golang`. 
</blockquote></details>

Good call. Done.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/17040#-:-KpI9uAOsZZEZhJVqNYq:b-o3baqt)*
<!-- Sent from Reviewable.io -->