@RaduBerinde 2017-07-18
Have you considered lifting the C implementations from the linux kernel instead? This kind of code is usually highly optimized there (in some cases, Intel contributes directly). I'm not sure if the GPL license is an issue.

https://github.com/torvalds/linux/blob/master/arch/x86/crypto/crc32-pclmul_asm.S