@benesch 2017-07-18
My understanding is we can't include or link any GPL-licensed code or else
we become a derived work and must license Cockroach under the GPL, too. (In
fact, I specifically *didn't* look at the Linux implementation.)

On Tue, Jul 18, 2017 at 8:57 AM RaduBerinde <notifications@github.com>
wrote:

> Have you considered lifting the C implementations from the linux kernel
> instead? This kind of code is usually highly optimized there (in some
> cases, Intel contributes directly). I'm not sure if the GPL license is an
> issue.
>
>
> https://github.com/torvalds/linux/blob/master/arch/x86/crypto/crc32-pclmul_asm.S
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/cockroachdb/cockroach/pull/17040#issuecomment-316055454>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AA15IHGxAl8B9Wunql5SSr3ugYITz8u8ks5sPKvGgaJpZM4OZgXu>
> .
>