@RaduBerinde 2017-07-18
LGTM. Haven't looked at the assembler in detail, but the tests are great (especially the cross-checks).

Too bad we're not using CRC32-C which is faster than CRC32-IEEE.