@benesch 2018-02-21
We're going to get rid of the CRC32C checksums entirely, so no need for this now!